==> 05-day5.Rmd <==
# Korrelation und Regression {#day5}

```{r include=FALSE}
knitr::opts_chunk$set(eval = FALSE)
```

## Lösung Tag 4

siehe Tag 4.

## Korrelation und Regression

https://figshare.com/articles/Storks_and_human_babies_data/839299/1 oder
https://github.com/jannikbuhr/dataIntro19/tree/master/data


```{r}
library(tidyverse)
library(broom)

storks <- read_csv("data/05_storks.csv")
storks
```

```{r}
ggplot(storks, aes(Storks, Birth)) +
  geom_point()
```

```{r}
# explain log transform and relevance in biology
```

```{r}
ggplot(storks, aes(Storks, Birth)) +
  geom_point() +
  scale_x_continuous(trans = "log10") +
  scale_y_continuous(trans = "log10") +
  annotation_logticks() +
  theme_classic()
```


```{r}
model <- lm(Birth ~ Storks, data = storks)
model
summary(model)
```

```{r}
ggplot(storks, aes(Storks, Birth)) +
  geom_point() +
  scale_x_continuous(trans = "log10", breaks = scales::log_breaks()) +
  scale_y_continuous(trans = "log10", breaks = scales::log_breaks()) +
  annotation_logticks() +
  theme_classic() +
  geom_smooth(method = "lm")
```


```{r}
model <- lm(log(Birth) ~ log(Storks), data = storks)
model
```

```{r}
summary(model)
```

```{r}
# explain correlation
# explain range of correlation coefficient
cor(storks$Storks, storks$Birth)
```

```{r}
cor(log(storks$Storks), log(storks$Birth))
```

```{r}
cor.test(storks$Storks, storks$Birth)
```

```{r}
cor(storks$Storks, storks$Birth, method = "spearman")
```

```{r}
model <- lm(storks$Birth ~ storks$Storks)
summary(model)
```

```{r}
cor(storks$Storks, storks$Birth)^2
```

```{r}
with(storks, {
  plot(Storks, Birth)
  abline(lm(Birth~Storks))
})
```

```{r}
with(storks, {
  plot(log(Storks), log(Birth))
  abline(lm(log(Birth)~log(Storks)))
})
```

```{r}
# show xkcd
# talk about confounding factors / variables
storks
```

## Tricks für viele Daten / Dateien

```{r}
paths <- dir("data", full.names = TRUE, pattern = ".csv")
map(paths, read_csv)
# map_dfr
```

## Non-Linear Regression

```{r}
Puromycin <- as_tibble(Puromycin)
Puromycin
```

```{r}
Puromycin %>% count(state)
```

```{r}
puro_treat <- Puromycin %>% 
  filter(state == "treated")
```

```{r}
plot(puro_treat$conc, puro_treat$rate)
```

```{r}
ggplot(puro_treat, aes(conc, rate)) + geom_point()
```

```{r}
michaelis_menten_fun <- function(conc, Vm, K) {
  (Vm * conc) / (K + conc)
}
```

```{r}
x = seq(0, 1, by = 0.001)
y = michaelis_menten_fun(x, Vm = 200, K = 0.2)
plot(x, y)
```


```{r}
plot(puro_treat$conc, puro_treat$rate)
curve(michaelis_menten_fun(conc = x, Vm = 200, K = 0.1),
      add = TRUE,
      col = "red",
      from = 0, to = 1,
      n = 100)
```

```{r}
model <- nls(rate ~ michaelis_menten_fun(conc, Vm, K),
             start = list(Vm = 200, K = 0.1),
             data  = puro_treat)
model
```

```{r}
coef(model)
```

```{r}
summary(model)$coefficients
```


```{r}
Vm_est <- coef(model)[1]
K_est  <- coef(model)[2]
```

```{r}
plot(puro_treat$conc, puro_treat$rate)
curve(michaelis_menten_fun(x, Vm = Vm_est, K = K_est),
      add = TRUE,
      col = "red")
```

```{r}
ggplot(puro_treat, aes(conc, rate)) +
  geom_point() +
  stat_function(fun = ~ michaelis_menten_fun(conc = .x, Vm = Vm_est, K = K_est),
                col = "red") +
  theme_classic()
```

```{r}
model2 <- nls(rate ~ SSmicmen(conc, Vm, K),
             data  = puro_treat)
model2
```

```{r}
predict(model2, newdata = list(conc = 1:10))
```



```{r}
nested_data <- Puromycin %>% 
  group_nest(state)
nested_data
```

```{r}
nested_data$data[[1]]
```


```{r}
all_models <- nested_data %>% 
  mutate(model = map(data, ~ nls(rate ~ michaelis_menten_fun(conc, Vm, K),
                                start = list(Vm = 200, K = 0.1),
                                data = .x))
  )

all_models
```

```{r}
fit_my_model <- function(df) {
  nls(rate ~ michaelis_menten_fun(conc, Vm, K),
                                start = list(Vm = 200, K = 0.1),
                                data = df)
}
```

```{r}
fit_my_model(puro_treat)
```


```{r}
all_models <- nested_data %>% 
  mutate(model = map(data, fit_my_model)
  )

all_models
```

```{r}
all_models$model[[1]]
```


```{r}
all_models <- all_models %>% 
  mutate(fitted = map(model, augment),
         coef   = map(model, coef))

all_models
```

```{r}
all_models %>% 
  unnest_wider(coef)
```

```{r}
all_models %>% 
  select(state, fitted) %>% 
  unnest(fitted)
```

```{r}
all_models %>% 
  select(state, fitted) %>% 
  unnest(fitted) %>% 
  ggplot(aes(conc, rate, color = state)) +
  geom_point() +
  geom_line(aes(y = .fitted))
```

```{r}
make_data <- function(model, from, to, n) {
  conc    = seq(from, to, length.out = n)
  .fitted = predict(model, newdata = list(conc = conc))
  tibble(conc, .fitted)
}

new_data <- all_models %>% 
  mutate(new_data = map(model, make_data, 0, 1, 100)) %>% 
  select(state, new_data) %>% 
  unnest(new_data)
```

```{r}
all_models %>% 
  select(state, fitted) %>% 
  unnest(fitted) %>% 
  ggplot(aes(conc, rate, color = state)) +
  geom_point() +
  geom_line(data = new_data, aes(y = .fitted))
```


## Übung

### Mit den Datasaurus Dozen Datensets

```{r}
datasauRus::datasaurus_dozen
```

- Visualisiere alle Sets gemeinsam in einem ggplot Scatterplot,
  nutze dazu `facet_wrap`.
- Füge mittels `geom_smooth` lineare Trendlinien hinzu


- Fitte eine Lineare Regression and jedes der Datensets. Nutze dazu die
  Techniken aus
  [R4DataScience: Many Models](https://r4ds.had.co.nz/many-models.html)
  und das `broom` package.
- Analysiere die Fits.
