==> 06-day6.Rmd <==

# Folgt {#day6}

```{r include=FALSE}
knitr::opts_chunk$set(eval = FALSE)
```

## Lösung Tag 5

```{r}
library(tidyverse)
library(broom)
```


### Mit den Datasaurus Dozen Datensets

```{r}
datasauRus::datasaurus_dozen
```

- Visualisiere alle Sets gemeinsam in einem ggplot Scatterplot,
  nutze dazu `facet_wrap`.
- Füge mittels `geom_smooth` lineare Trendlinien hinzu


- Fitte eine Lineare Regression and jedes der Datensets. Nutze dazu die
  Techniken aus
  [R4DataScience: Many Models](https://r4ds.had.co.nz/many-models.html)
  und das `broom` package.
- Analysiere die Fits.

```{r}
dinos <- datasauRus::datasaurus_dozen
dinos
```

```{r}
dinos %>% 
  count(dataset)
```

```{r}
dinos %>% 
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_wrap("dataset", scales = "free_y") +
  ggthemes::theme_few()
```

```{r}
nested_dinos <- dinos %>% 
  group_nest(dataset)
nested_dinos
```

```{r}
beispiel <- nested_dinos$data[[1]]

model <- lm(y ~ x, data = beispiel)
summary(model)

glance(model)
tidy(model)
coef(model)
```


```{r}
all_models <- nested_dinos %>% 
  mutate(
    model        = map(data, ~ lm(y ~ x, data = .x)),
    glance       = map(model, glance),
    coefficients = map(model, coef),
    corr         = map(data, ~ cor(.x$x, .x$y))
  )
all_models
```

```{r}
all_models %>% 
  select(dataset, coefficients) %>% 
  unnest_wider(coefficients)
```

```{r}
all_models %>% 
  select(dataset, corr) %>% 
  unnest_wider(corr)
```

## Datenanalyse, Beispiel: Global Plastic Waste




## Feedbackrunde
